+++ nsw_fuel/const.py
@@ -2,8 +2,6 @@
 
 from logging import Logger, getLogger
 
-LOGGER: Logger = getLogger(__package__)
-
 AUTH_URL = "https://api.onegov.nsw.gov.au/oauth/client_credential/accesstoken?grant_type=client_credentials"
 BASE_URL = "https://api.onegov.nsw.gov.au"
 REFERENCE_ENDPOINT = "/FuelCheckRefData/v2/fuel/lovs"
